Refactor escape impls between two modules.
